requirements.txt
2captcha-python==1.2.7
attrs==23.2.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
h11==0.14.0
idna==3.7
outcome==1.3.0.post0
pycparser==2.22
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.22.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
XlsxWriter==3.2.0
